refactor: table added, fix: start over button functionality
